Break controls into components
